+++ tests/dummy/app/controllers/single-select.js
@@ -8,13 +8,16 @@ export default Controller.extend({
     this.set('bloomer', { id: 13, name: 'Bloomer' });
     this.set('brioche', { id: 14, name: 'Brioche' });
   },
+
   actions: {
     selectedBread(bread) {
       this.set('selectedBread', bread);
     },
+
     action1() {
       this.set('action1', true);
     },
+
     action2() {
       this.set('action2', true);
     }